+++ api-todo/src/data/todo.validation.ts
@@ -1,7 +1,6 @@
 import Joi from 'joi';
 import initialData from './todo.initial.json';
 import { Todo } from './todo.types';
-import { toError } from '../utils/error';
 
 export const INITIAL_PARTIAL_DATA: Partial<Todo>[] = initialData;
 
@@ -10,24 +9,16 @@ export { initialData };
 export const MAX_LENGTH_TITLE = 50;
 export const MAX_LENGTH_DESCRIPTION = 500;
 
-export const todoSchema = Joi.object({
-  id: Joi.alternatives().try(Joi.number().greater(0), Joi.string()).required(),
-  title: Joi.string().min(1).max(MAX_LENGTH_TITLE).required(),
-  description: Joi.string().min(1).max(MAX_LENGTH_DESCRIPTION).required(),
-  createdAt: Joi.date().iso().required(),
-  updatedAt: Joi.date().iso().allow(null),
-}).unknown(false);
-
 // Title is only required field
 export const updateTodoSchema = Joi.object({
   id: Joi.alternatives().try(Joi.number().greater(0), Joi.string()),
   title: Joi.string().min(1).max(MAX_LENGTH_TITLE).required(),
   description: Joi.string().min(1).max(MAX_LENGTH_DESCRIPTION),
-  createdAt: Joi.date().iso(),
-  updatedAt: Joi.date().iso(),
+  createdAt: Joi.date().iso().required(),
+  updatedAt: Joi.date().iso().allow(null), // FIXED: initial update won't have updatedAt value
 }).unknown(false);
 
-export const todoPartialSchema = Joi.object({
+export const newTodoSchema = Joi.object({
   title: Joi.string().min(1).max(MAX_LENGTH_TITLE).required(),
   description: Joi.string().min(1).max(MAX_LENGTH_DESCRIPTION),
 }).unknown(false);
@@ -39,29 +30,30 @@ export interface TodoValidation {
 }
 
 export const isValidPartial = (todo: Partial<Todo>): TodoValidation => {
-  const { error } = updateTodoSchema.validate(todo);
-  if (error) {
-    if (error.details) {
-      const messages = error.details.map((item) => item.message);
-      return {
-        valid: false,
-        error: toError(messages.join(', ')),
-        todo: null,
-      };
+    const { error } = newTodoSchema.validate(todo);
+    if (error) {
+      if (error.details) {
+        const messages = error.details.map((item) => item.message).join(', ');
+        const returnObj = {
+          valid: false,
+          error: new Error(`todo.validation failed:isValidPartial ${messages}`),
+          todo: null,
+        };
+        return returnObj;
+      }
     }
-  }
-  return { valid: true, error: null, todo };
+    return { valid: true, error: null, todo };
 };
 
 export const isValidAll = (todo: Todo): TodoValidation => {
-  const { error } = todoSchema.validate(todo);
+  const { error } = updateTodoSchema.validate(todo);
   if (error) {
     // if error is a ValidationErrorItem, join all messages together
     if (error.details) {
       const messages = error.details.map((item) => item.message);
       return {
         valid: false,
-        error: toError(messages.join(', ')),
+        error: new Error(`todo.validation failed:isValidAll ${messages}`),
         todo: null,
       };
     }