+++ client-todo/src/todo/index.tsx
@@ -1,58 +1,30 @@
-import { useState } from 'react';
-import useSWR, { mutate } from 'swr';
 import TodoForm from './components/form';
+import { useState } from 'react';
 import List from './components/list';
-import { NewTodo, Todo } from './models';
-import { addTodo, deleteTodo, API_ADD_TODO, API_DELETE_TODO, API_GET_ALL_TODOS } from './service';
-import { fetcher } from './api';
+import { Todo } from './models';
+import { useTodos } from './useTodo';
 
 export default function Todo() {
-    const [requestError, setRequestError] = useState('');
-    const { data, error, isLoading } = useSWR(API_GET_ALL_TODOS, fetcher)
+    const [currentTodo, setCurrentTodo] = useState<Todo | undefined>(undefined);
+    const { data, error, isLoading, isValidating, addTodo, updateTodo, removeTodo } = useTodos();
 
-    async function handleSubmit(newTodoItem: NewTodo) {
-        setRequestError('');
-
-        try {
-            const result = await addTodo(newTodoItem);
-
-            if (!result.ok) throw new Error(`result: ${result.status} ${result.statusText}`);
-            const { data: savedTodo, error: returnedError } = await result.json();
-
-            if (returnedError) throw new Error(returnedError);
-
-            mutate(API_ADD_TODO, [...data, savedTodo], false);
-
-        } catch (error: unknown) {
-            setRequestError(String(error));
-        }
-    }
-
-    async function handleDelete(id: number) {
-        setRequestError('');
-        try {
-            const result = await deleteTodo(id);
-            if (!result.ok) throw new Error(`result: ${result.status} ${result.statusText}`);
-            const { data: deletedTodo, error: returnedError } = await result.json();
-
-            if (returnedError) throw new Error(returnedError);
-
-            mutate(API_DELETE_TODO, data.filter((todo: Todo) => todo.id !== deletedTodo.id), false);
-        } catch (error: unknown) {
-            setRequestError(String(error));
-        }
-    }
-
-    if (error || requestError) return <div >failed to load {error ? JSON.stringify(error) : requestError}</div>
+    if (error) return <div >failed to load {JSON.stringify(error)}</div>
     if (!error && isLoading) return <div >loading...{JSON.stringify(isLoading)}</div>
+    if (!error && isValidating) return <div >isValidating...{JSON.stringify(isValidating)}</div>
+
+    const setThisTodo = (todo: Todo) => {
+        console.log('index setCurrentTodo', todo)
+        setCurrentTodo(todo);
+    }
+
 
     return (
         <div >
-            <TodoForm onSubmit={handleSubmit} requestError={requestError} />
+            <TodoForm newSubmit={addTodo} setCurrentTodo={setCurrentTodo} updateSubmit={updateTodo} todoToEdit={currentTodo} requestError={error} />
             <div >
                 { data!==undefined 
                     && data.length>0 
-                    && <List todos={data} onDelete={handleDelete} />
+                    && <List todos={data} onDelete={removeTodo} setCurrentTodo={setThisTodo} />
                 }
             </div>
         </div>